WebSellingShoes/ChatbotAPI/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pyodbc
pydantic